tools: split linters into tools/lint/checks
Each linter becomes its own executable under tools/lint/checks, driven
by harness.py, and its own prek hook gated on the files it cares about.
The in-code suffix guards move into each hook's files/exclude patterns,
so prek skips a check's process entirely when nothing relevant changed.

Gameflow schema validation now compiles the schema with fastjsonschema
and runs only when a gameflow file or the schema itself changes, rather
than on every commit.

Retire tools/additional_lint and tools/shared/linting.py. The tools test
helper stubbed jsonschema for an import path that no longer exists; drop
it, keeping the pyjson5 stub json_utils still needs.
