+++ src/tests/unit/tools_helper.py
@@ -16,12 +16,11 @@ from pathlib import Path
 ROOT = Path(__file__).resolve().parents[3]
 sys.path.insert(0, str(ROOT / "tools"))
 
-# update_game_strings reaches jsonschema through shared.linting, but only inside
-# a validation step unrelated to anything tested here. Satisfy the import rather
-# than skip the tests: a skipped test proves nothing.
-for _optional in ("jsonschema", "pyjson5"):
-    if importlib.util.find_spec(_optional) is None:
-        sys.modules[_optional] = types.ModuleType(_optional)
+# update_game_strings parses JSON5 through shared.json_utils, which imports
+# pyjson5 - unrelated to anything tested here. Satisfy the import rather than
+# skip the tests: a skipped test proves nothing.
+if importlib.util.find_spec("pyjson5") is None:
+    sys.modules["pyjson5"] = types.ModuleType("pyjson5")
 
 
 def load(name: str):
