+++ tools/lint/checks/newlines
@@ -1,0 +1,19 @@
+#!/usr/bin/env python3
+import sys
+from pathlib import Path
+
+sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
+
+from harness import LintWarning, file_check
+
+
+def check(path, text):
+    if not text:
+        return
+    if not text.endswith("\n"):
+        yield LintWarning(path, "missing newline character at end of file")
+    if text.endswith("\n\n"):
+        yield LintWarning(path, "extra newline character at end of file")
+
+
+file_check(check)
